split up too generic helper file
